+++ resources/src/test/kotlin/io/bkbn/kompendium/resources/KompendiumResourcesTest.kt
@@ -14,9 +14,11 @@ import io.ktor.server.application.install
 import io.ktor.server.resources.Resources
 import io.ktor.server.resources.get
 import io.ktor.server.response.respondText
+import io.ktor.server.routing.Route
+import io.ktor.server.routing.route
 
 class KompendiumResourcesTest : DescribeSpec({
-  describe("Resource Tests") {
+  describe("NotarizedResources Tests") {
     it("Can notarize a simple resource") {
       openApiTestAllSerializers(
         snapshotName = "T0001__simple_resource.json",
@@ -117,4 +119,72 @@ class KompendiumResourcesTest : DescribeSpec({
       }
     }
   }
+  describe("NotarizedResource Tests") {
+    it("Can notarize resources in route") {
+      openApiTestAllSerializers(
+        snapshotName = "T0003__resources_in_route.json",
+        applicationSetup = {
+          install(Resources)
+        }
+      ) {
+        route("/api") {
+          typeEditDocumentation()
+          get<Type.Edit> { edit ->
+            call.respondText("Listing ${edit.parent.name}")
+          }
+          typeOtherDocumentation()
+          get<Type.Other> { other ->
+            call.respondText("Listing ${other.parent.name}, page ${other.page}")
+          }
+        }
+      }
+    }
+  }
 })
+
+private fun Route.typeOtherDocumentation() {
+  install(NotarizedResource<Type.Other>()) {
+    parameters = listOf(
+      Parameter(
+        name = "name",
+        `in` = Parameter.Location.path,
+        schema = TypeDefinition.STRING
+      ),
+      Parameter(
+        name = "page",
+        `in` = Parameter.Location.path,
+        schema = TypeDefinition.INT
+      )
+    )
+    get = GetInfo.builder {
+      summary("Other")
+      description("example resource")
+      response {
+        responseCode(HttpStatusCode.OK)
+        responseType<TestResponse>()
+        description("does great things")
+      }
+    }
+  }
+}
+
+private fun Route.typeEditDocumentation() {
+  install(NotarizedResource<Type.Edit>()) {
+    parameters = listOf(
+      Parameter(
+        name = "name",
+        `in` = Parameter.Location.path,
+        schema = TypeDefinition.STRING
+      )
+    )
+    get = GetInfo.builder {
+      summary("Edit")
+      description("example resource")
+      response {
+        responseCode(HttpStatusCode.OK)
+        responseType<TestResponse>()
+        description("does great things")
+      }
+    }
+  }
+}